main/serializers.py
from rest_framework import serializers
from .models import ShortenedURL


class ShortenedURLSerializer(serializers.ModelSerializer):
    class Meta:
        model = ShortenedURL
        fields = ('id', 'original_url', 'shortened_url', 'clicks', 'created_at')

